requirements.txt
transformers
pytorch-lightning
wandb~=0.10.15
nltk~=3.5
pyyaml~=5.3.1
tqdm~=4.56.0
sklearn
lemminflect